added changelog entry moved cache readme tests to test_cacheprovider.py
