+++ changelog/3519.feature.rst
@@ -1,0 +1,1 @@
+now `.pytest_cache` is created with a README.md to indicate it's usage,options etc.
